backend/requirements.txt
fastapi
uvicorn
python-dotenv
psycopg2-binary
sqlalchemy
python-jose[cryptography]
httpx
pydantic
